requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap4==23.2
django-filter==23.2
django-tables2==2.6.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
bootstrap4==0.1.0
gunicorn==20.1.0
whitenoise==6.5.0
